+++ src/App.jsx
@@ -1,21 +1,20 @@
 // App.js
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Layout from './components/layout';
-import Home from './components/home';
-import StudentR from './components/Registration_forms/student';
-import UniversityR from './components/Registration_forms/university';
-import OPartyR from './components/Registration_forms/OParty';
-import StudentH from "./components/Home/student";
-import UniversityH from "./components/Home/university";
+import Home from './pages/home';
+import StudentR from './pages/Registration_forms/student';
+import UniversityR from './pages/Registration_forms/university';
+import OPartyR from './pages/Registration_forms/OParty';
+import StudentH from "./pages/Home/student";
+import UniversityH from "./pages/Home/university";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/js/bootstrap.min.js';
 import 'jquery/dist/jquery.min.js';
 import 'popper.js/dist/umd/popper.min.js'
 import $ from 'jquery';
-import Records from './components/records';
-import Notification from './components/notification';
-
+import Records from './components/Urecords';
+import Notification from './components/Unotification';
+import OpartyH from "./pages/Home/OParty"
 function App() {
   
   return (
