+++ src/components/Unotification.jsx
@@ -13,7 +13,7 @@ import {
   MDBBtn
 } from "mdb-react-ui-kit";
 import Footer from "./footer";
-import Sidebarcomponent from "./sidebar";
+import Sidebarcomponent from "./Usidebar";
 
 export default function Notification() {
   return (
