+++ src/pages/home.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Layout from './layout';
-import Footer from './footer';
+import Layout from '../components/layout';
+import Footer from '../components/footer';
 export default function home() {
   return (
     <div>
