+++ src/styles/Home/OParty.css
@@ -1,0 +1,20 @@
+.app-container {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+  }
+  
+  .content-wrapper {
+    display: flex;
+    flex: 1;
+    flex-grow: 1;
+  }
+  
+  .bg-image {
+    flex: 1;
+    background-size: cover;
+    position: relative;
+    overflow: hidden;
+  }
+  
+
